Move entity reference resolution out of parse.ts
parse.ts mixed two concerns: pure conversion of attribute literals into
engine values, and a DOM-aware entity reference subsystem - safe
document and inclusive-scope queries, the entity-fronting kinds and
selectors, lexical name lookup, entity extraction, unresolved-cause
wording, the bare-id hint, and the reporting resolveEntity wrapper.

The reference subsystem moves verbatim to an internal
entity-reference.ts, leaving parse.ts with only literal parsing - it no
longer imports the engine Entity type or touches the DOM. Consumers
split their imports by concern: the entity-accepting components take
resolveEntity from the new module beside their literal parsers, and
pc-script takes the lower-level lookup and warning helpers.

The reference grammar, lookup order, and warning text are unchanged;
the integration suite that pins them moves with the boundary, renamed
from get-entity.test.ts to entity-reference.test.ts, while the pure
parser tests stay in the unit tier. The new module is excluded from
the Custom Elements Manifest like the other non-element modules, and
its emitted declaration strips to the module comment - the manifest,
editor integrations and public declarations are byte-identical.

Fixes #442
